refactor(api): merge agent module into single file
Merge agent/contracts.ts, agent/runner.ts, and agent/service.ts into
a single lib/agent.ts. Move agent/tools.ts to lib/agent-tools.ts.
Eliminates the agent/ directory and removes one intermediate type
(AgentRunnerChunk as a cross-module handoff) while preserving the
createRunner testability seam.
