+++ apps/api/src/lib/agent-tools.ts
